Revert "Refactor/refactoring basics"
