+++ README.md
@@ -1,50 +1,3 @@
 # Python_Pilgrim
 
-## Overview
-
-**Python_Pilgrim** is a Python-based travel diary application designed to help users document and manage their travel experiences. The project provides tools for recording trips, organizing travel notes, and storing memories in a structured and accessible format.
-
-## Features
-
-- Create and manage travel diaries
-- Add, edit, and delete travel entries
-- Organize trips by date, location, or theme
-- Store photos, notes, and other media
-- Export and share travel logs
-
-## Requirements
-- Python 3.8 or higher
-- Linux operating system (tested on Ubuntu 20.04+)
-- Visual Studio Code (VSCode) for development (optional but strongly recommended)
-- pip (Python package installer)
-- Optional: virtualenv for isolated environments
-
-## Installation
-
-1. Clone the repository:
-    ```bash
-    git clone https://github.com/gmbrax/Pilgrim.git
-    ```
-2. Navigate to the project directory:
-    ```bash
-    cd Pilgrim
-    ```
-3. Create a virtual environment and, then, activate it:
-    ```bash
-    python -m venv .venv
-    source .venv/bin/activate
-    ```
-4. Install the required dependencies:
-    ```bash
-    pip install -r requirements.txt
-    ```
-
-## Usage
-
-To run the main application, execute:
-
-```bash
-python <???>.py
-```
-
-This will start the Python_Pilgrim application. Follow the on-screen instructions to create and manage your travel diaries.
+Python Based Travel Diary
