--- requirements.txt
@@ -1,3 +1,0 @@
-greenlet==3.2.3
-SQLAlchemy==2.0.41
-typing_extensions==4.14.0
