--- src/models/photo_in_entry.py
@@ -1,8 +1,0 @@
-from sqlalchemy import Table, Column, Integer, ForeignKey
-
-from src.database import Base
-
-photo_entry_association = Table('photo_entry_association', Base.metadata,
-Column('id', Integer, primary_key=True, autoincrement=True),
-    Column('fk_photo_id', Integer, ForeignKey('photos.id'),nullable=False),
-    Column('fk_entry_id', Integer, ForeignKey('entries.id'),nullable=False))
